+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - **New**: TypeScript added in node, shared files between node and app
 - **Fix**: Swear words have been removed from Italian winning words
 - **New**: All langs are using FrequencyWords as part of building winning dictionary
+- **New**: A hint was added informing that yesterday's word could be wrong due to the update
 
 #### Version 3.21 (12.05.2024)
 - **New**: Italian game language added